Improved linux setup

+++ linux/setup.sh
@@ -12,7 +12,6 @@ function install_homebrew() {
   if [ ! -f "`which brew`" ]; then
     printf "${CYAN}############################################################## Installing homebrew${NC}\n"
     NONINTERACTIVE=1 /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
-    echo 'eval "$(/home/linuxbrew/.linuxbrew/bin/brew shellenv)"' >> ~/.zshrc
     eval "$(/home/linuxbrew/.linuxbrew/bin/brew shellenv)"
     install build-essential
   fi
@@ -20,6 +19,7 @@ function install_homebrew() {
 }
 
 function install_packages() {
+  printf "${CYAN}############################################################## Installing Homebrew/Apt packages${NC}\n"
   install 'wget'
   install 'glances'
   install 'watch'
@@ -48,7 +48,8 @@ function install_packages() {
 
 function setup_aliases() {
   printf "${CYAN}############################################################## Setting up aliases......\n${NC}"
-  echo "source $basedir/base/aliases" > ~/.aliases
+  add_to_file_unique "source $basedir/base/aliases" ~/.aliases
+  add_to_file_unique "source $basedir/linux/aliases" ~/.aliases
 }
 
 function setup_git() {
@@ -71,7 +72,9 @@ function install_node() {
   install 'gnupg'
   sudo mkdir -p /etc/apt/keyrings
   NODE_MAJOR=21
-  curl -fsSL https://deb.nodesource.com/gpgkey/nodesource-repo.gpg.key | sudo gpg --dearmor -o /etc/apt/keyrings/nodesource.gpg
+  if [ ! -f /etc/apt/keyrings/nodesource.gpg ] ; then
+    curl -fsSL https://deb.nodesource.com/gpgkey/nodesource-repo.gpg.key | sudo gpg --dearmor -o /etc/apt/keyrings/nodesource.gpg
+  fi
   echo "deb [signed-by=/etc/apt/keyrings/nodesource.gpg] https://deb.nodesource.com/node_$NODE_MAJOR.x nodistro main" | sudo tee /etc/apt/sources.list.d/nodesource.list
   sudo apt-get update
   install 'nodejs' # required by coc.vim
@@ -82,8 +85,8 @@ function setup_vim() {
   printf "${CYAN}############################################################## Setting up vim......\n${NC}"
   install 'vim'
   ln -sf $basedir/base/.vim ~/
+  printf "${CYAN}Ignore the error saying ' Cannot find color scheme'. Just hit Enter\n${NC}"
   curl -fLo ~/.vim/autoload/plug.vim --create-dirs https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
-  echo "${CYAN}Ignore the error saying ' Cannot find color scheme'. Just hit Enter${NC}"
   vi +'PlugInstall' +qa
 }
 
@@ -96,11 +99,9 @@ function setup_zsh() {
   if [ ! -f $HOME/antigen.zsh ] ; then
     curl -L git.io/antigen > $HOME/antigen.zsh
   fi
-  echo "source $basedir/zsh/zshrc" >> ~/.zshrc
-  echo "alias clip=clip.exe" >> ~/.zshrc # like pbcopy on windows
+  add_to_file_unique "source $basedir/zsh/zshrc" ~/.zshrc
   touch ~/.zsh_history # for mcfly
-  echo "source /home/raziel/dotfiles/zsh/functions.zsh" >> ~/.zshrc # replace with `realpath functions.zsh`
-  echo "eval \"$(/home/linuxbrew/.linuxbrew/bin/brew shellenv)\"" >> ~/.zshrc # replace hardcoded path
+  eval $(/home/linuxbrew/.linuxbrew/bin/brew shellenv)
   binstall 'jandedobbeleer/oh-my-posh/oh-my-posh' # https://ohmyposh.dev/
   /bin/zsh -i -c "source ~/antigen.zsh"
 }
@@ -109,15 +110,12 @@ function setup_tmux() {
   printf "${CYAN}############################################################## Setting up tmux......\n${NC}"
   install 'tmux'
   install 'xdg-utils' # needed by tmux-open
-  ln -sf $basedir/base/.tmux.conf ~/
-  echo "source-file ~/.tmux.conf" > ~/.tmate.conf
-  if [ ! -d ~/.tmux/plugins/tpm ] ; then
-    git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
-    ~/.tmux/plugins/tpm/bin/install_plugins
-  fi
-  tmux source ~/.tmux.conf
+
+  checkout_or_update_git_repo https://github.com/gpakosz/.tmux.git ~/.tmux master
+  ln -s -f .tmux/.tmux.conf ~/
+  ln -sf $basedir/base/.tmux.conf.local ~/
+  add_to_file_unique "source-file ~/.tmux.conf" ~/.tmate.conf
 }
-printf "${CYAN}############################################################## Installing Homebrew/Apt packages${NC}\n"
 
 update_apt
 install_homebrew
@@ -130,7 +128,7 @@ setup_zsh
 setup_tmux
 cleanup
 
-printf "${CYAN}############################################################## Linux Setup complete! \n${NC}"
 printf "${CYAN}############################################################## Switching to ZSH \n${NC}"
 
 chsh -s `which zsh`
+printf "${CYAN}############################################################## Linux Setup complete! \n${NC}"
